% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_by_strata_by_trt.R
\name{crit_bb_nsubev_01}
\alias{crit_bb_nsubev_01}
\title{Evaluate inclusion criterion for by-strata-by-treatment statistics based on the number of subjects with events in the total population.}
\usage{
crit_bb_nsubev_01(dat, event_index, subjectid_var, n_subj_event_min, ...)
}
\arguments{
\item{dat}{data.table. The analysis data set.}

\item{event_index}{vector of integers that index the rows in \code{dat} that match
the definition of an 'event'. Matching is done via the \code{INDEX_} column in
\code{dat}.}

\item{subjectid_var}{character. Name of the subject identifier variable in \code{dat} (default is "USUBJID").}

\item{n_subj_event_min}{The minimum number of subjects with events required in the total population.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A Boolean value indicating whether the criterion is met.
}
\description{
This function checks if the number of subjects with events in the total population is greater than or equal to a specified minimum.
}
\examples{
dat <- data.table::data.table(USUBJID = c("S1", "S2", "S3", "S4"))
dat[, INDEX_ := .I]
data.table::setkey(dat, INDEX_)
# Returns TRUE: 2 subjects with events >= minimum of 2
crit_bb_nsubev_01(dat, event_index = c(1L, 2L),
                  subjectid_var = "USUBJID", n_subj_event_min = 2L)
# Returns FALSE: 1 subject with events < minimum of 2
crit_bb_nsubev_01(dat, event_index = c(1L),
                  subjectid_var = "USUBJID", n_subj_event_min = 2L)
}
