% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_by_strata_by_trt.R
\name{crit_bb_pval_01}
\alias{crit_bb_pval_01}
\title{Evaluate inclusion criterion for by-strata-by-treatment statistics based on the p-value in the total population.}
\usage{
crit_bb_pval_01(
  dat,
  event_index,
  cell_index,
  treatment_var,
  treatment_refval,
  subjectid_var,
  pval_max,
  ...
)
}
\arguments{
\item{dat}{data.table. The analysis data set.}

\item{event_index}{vector of integers that index the rows in \code{dat} that match
the definition of an 'event'. Matching is done via the \code{INDEX_} column in
\code{dat}.}

\item{cell_index}{A vector of integers referencing the rows of \code{dat} (as
specified by the \code{INDEX_} column in \code{dat}) that match the population to be
analyzed. See the "Endpoint Events" vignette in {ramnog}
for more information.}

\item{treatment_var}{character. The name of the treatment variable in \code{dat}.}

\item{treatment_refval}{character. The reference value of the treatment variable in \code{dat}.}

\item{subjectid_var}{character. Name of the subject identifier variable in \code{dat} (default is "USUBJID").}

\item{pval_max}{The maximum p-value.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A Boolean value indicating whether the criterion is met.
}
\description{
This function checks if the p-value in the total population is below a specified maximum.
}
