% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_endpoint.R
\name{crit_ep_nsubev_01}
\alias{crit_ep_nsubev_01}
\title{Evaluate endpoint criterion based on number of subjects with events.}
\usage{
crit_ep_nsubev_01(
  dat,
  event_index,
  subjectid_var,
  treatment_var,
  n_subj_event_min,
  ...
)
}
\arguments{
\item{dat}{data.table. The analysis data set.}

\item{event_index}{vector of integers that index the rows in \code{dat} that match
the definition of an 'event'. Matching is done via the \code{INDEX_} column in
\code{dat}.}

\item{subjectid_var}{character. Name of the subject identifier variable in \code{dat} (default is "USUBJID").}

\item{treatment_var}{The name of the variable in dat containing the treatment group assignments.}

\item{n_subj_event_min}{The minimum number of subjects with events required in at least one study arm.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A Boolean value indicating whether the criterion is met.
}
\description{
This function checks if the number of subjects with events in at least one study arm is greater than or equal to a specified minimum.
}
